# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AdversaryFlow API
  version: 0.3.0
  description: Local API for ATT&CK actors and guided emulation plans. When the service is bound beyond loopback, every API operation requires BearerToken authentication.
servers:
  - url: http://127.0.0.1:5000
paths:
  /api/session:
    get:
      summary: Issue a same-origin token for mutating local operations
      responses:
        "200":
          description: Browser session token
          content:
            application/json:
              schema:
                type: object
                required: [csrf_token, version]
                properties:
                  csrf_token: { type: string }
                  version: { type: string }
  /api/bootstrap:
    get:
      summary: Report background ATT&CK bootstrap progress
      responses:
        "200": { description: Data is ready }
        "202": { description: Bootstrap is running }
        "503": { description: Bootstrap has not started or failed }
    post:
      summary: Start or retry background ATT&CK bootstrap
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200": { description: Data is ready }
        "202": { description: Bootstrap started or is already running }
        "403": { $ref: "#/components/responses/Error" }
  /api/health:
    get:
      summary: Report liveness and ATT&CK readiness
      responses:
        "200":
          description: ATT&CK data is ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
        "503":
          description: Process is live but ATT&CK data is not ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Health" }
  /api/actors:
    get:
      summary: List groups and campaigns with mapped techniques
      parameters:
        - $ref: "#/components/parameters/Domains"
      responses:
        "200":
          description: Actor collection and data version
          content:
            application/json:
              schema:
                type: object
                required: [actors, domains, data_version, version]
                properties:
                  actors: { type: array, items: { $ref: "#/components/schemas/Actor" } }
                  domains: { type: array, items: { type: string } }
                  data_version: { type: string }
                  version: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /api/workflow/{stix_id}:
    get:
      summary: Build a versioned workflow for one actor
      parameters:
        - name: stix_id
          in: path
          required: true
          schema: { type: string }
        - $ref: "#/components/parameters/Domains"
      responses:
        "200":
          description: Ordered workflow with commands and metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "503": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
  /api/execution-kit:
    post:
      summary: Generate an offline CSV and PowerShell/Bash operator handoff
      description: Serializes a schema 2.0 browser plan into a ZIP. The service does not execute commands.
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [schema_version, tool, actor, scope, execution_context, stages]
              properties:
                schema_version: { const: "2.0" }
                tool: { const: AdversaryFlow }
                actor: { type: object }
                scope: { type: object }
                execution_context: { type: object }
                stages: { type: array, minItems: 1, maxItems: 32 }
      responses:
        "200":
          description: ZIP containing the operator CSV and matching standalone runner
          content:
            application/zip:
              schema: { type: string, format: binary }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "413": { $ref: "#/components/responses/Error" }
  /api/refresh:
    post:
      summary: Refresh selected ATT&CK domains and rebuild their index
      parameters:
        - $ref: "#/components/parameters/Domains"
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Refreshed data version
          content:
            application/json:
              schema:
                type: object
                required: [status, domains, data_version]
                properties:
                  status: { const: refreshed }
                  domains: { type: array, items: { type: string } }
                  data_version: { type: string }
                  cache: { type: object }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
        "429": { $ref: "#/components/responses/Error" }
        "500": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    BearerToken:
      type: http
      scheme: bearer
      description: Required on every API route only when non-loopback remote mode is enabled.
  parameters:
    CsrfToken:
      name: X-AdversaryFlow-CSRF
      in: header
      required: true
      description: Same-origin token returned by GET /api/session.
      schema: { type: string }
    Domains:
      name: domains
      in: query
      description: Comma-separated ATT&CK domains.
      schema:
        type: string
        default: enterprise
        examples: [enterprise, "enterprise,ics,mobile"]
  schemas:
    Actor:
      type: object
      required: [stix_id, attack_id, name, type, aliases, description, technique_count]
      properties:
        stix_id: { type: string }
        attack_id: { type: string }
        name: { type: string }
        type: { enum: [group, campaign] }
        aliases: { type: array, items: { type: string } }
        description: { type: string }
        technique_count: { type: integer, minimum: 0 }
      additionalProperties: false
    Workflow:
      type: object
      required: [actor, summary, kill_chain, stages, metadata]
      properties:
        actor: { type: object }
        summary:
          type: object
          required: [total_techniques, unique_stages, curated_commands, fallback_commands]
        kill_chain:
          type: array
          items:
            type: object
            required: [tactic, title]
            properties:
              tactic: { type: string }
              title: { type: string }
        stages:
          type: array
          items: { $ref: "#/components/schemas/Stage" }
        metadata:
          type: object
          required: [domains, data_version, version]
          properties:
            domains: { type: array, items: { type: string } }
            data_version: { type: string }
            version: { type: string }
    Stage:
      type: object
      required: [tactic, title, techniques]
      properties:
        tactic: { type: string }
        title: { type: string }
        techniques:
          type: array
          items: { $ref: "#/components/schemas/Technique" }
    Technique:
      type: object
      required: [attack_id, name, commands, command_source]
      properties:
        attack_id: { type: string }
        name: { type: string }
        stix_id: { type: string }
        description: { type: string }
        tactics: { type: array, items: { type: string } }
        platforms: { type: array, items: { type: string } }
        is_subtechnique: { type: boolean }
        data_sources: { type: array, items: { type: string } }
        detection: { type: string }
        url: { type: [string, "null"], format: uri }
        commands:
          type: array
          items: { $ref: "#/components/schemas/Command" }
        command_source: { enum: [curated, fallback] }
    Command:
      type: object
      required: [platform, command, note, cleanup, risk, side_effects, requires_admin, requires_network, network_targets, prerequisites, expected_telemetry, expected_output, timeout_seconds, rollback, cleanup_required, acknowledgment_required]
      properties:
        platform: { type: string }
        command: { type: string }
        note: { type: string }
        cleanup: { type: string }
        risk: { enum: [low, medium, high] }
        side_effects: { type: array, items: { type: string } }
        requires_admin: { type: boolean }
        requires_network: { type: boolean }
        network_targets: { type: array, items: { type: string } }
        prerequisites: { type: array, items: { type: string } }
        expected_telemetry: { type: string }
        expected_output: { type: string }
        timeout_seconds: { type: integer }
        rollback: { type: string }
        cleanup_required: { type: boolean }
        acknowledgment_required: { type: boolean }
        exercise_kind: { enum: [technique_relevant_bounded] }
        fidelity: { enum: [bounded_synthetic] }
        evidence_source: { enum: [self_reported_receipt] }
        unsupported: { type: boolean }
        restricted: { type: boolean }
        telemetry_acceptance:
          type: object
          required: [technique_id, scenario, activity_event_types, minimum_activity_events, requirements, limitation]
          properties:
            technique_id: { type: string }
            scenario: { type: string }
            activity_event_types: { type: array, minItems: 1, items: { type: string } }
            minimum_activity_events: { type: integer, minimum: 1 }
            requirements: { type: array, minItems: 1, items: { type: string } }
            limitation: { type: string }
          additionalProperties: false
      additionalProperties: false
    Health:
      type: object
      required: [status, ready, loading, phase, version, error, attack_data, service]
      properties:
        status: { enum: [ready, degraded] }
        ready: { type: boolean }
        loading: { type: boolean }
        phase: { enum: [not_started, loading, refreshing, ready, failed] }
        version: { type: string }
        error: { type: [string, "null"] }
        attack_data: { type: object }
        service: { type: object }
    Error:
      type: object
      required: [error, message, version]
      properties:
        error: { type: string }
        message: { type: string }
        request_id: { type: string }
        version: { type: string }
  responses:
    Error:
      description: Structured application error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }